inventory_system/backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
pandas==2.1.4
python-dotenv==1.0.0
python-socketio==5.10.0
eventlet==0.33.3
